requirements.txt
# Fishtown Web Design Social Media Automation - Requirements
# Core dependencies for content generation and social media posting

# OpenAI API for AI-powered content generation
openai==1.12.0

# Environment variable management
python-dotenv>=1.0.0,<2.0.0

# Task scheduling for automated daily posting
schedule>=1.2.0,<2.0.0

# Image processing for generated images
Pillow>=10.0.1,<11.0.0

# HTTP requests for API calls and image downloads
requests==2.31.0
httpx==0.27.0

# Facebook Graph API integration
facebook-sdk>=3.1.0,<4.0.0

# Instagram API integration (currently disabled but available)
instagram-private-api>=1.6.0,<2.0.0

# Additional utilities that may be needed
# JSON handling (built-in, but listed for clarity)
# datetime (built-in)
# hashlib (built-in)
# os (built-in)
# sys (built-in)
# time (built-in)
# random (built-in)
# typing (built-in)
# io (built-in)

# Optional: For better error handling and logging
# logging (built-in)

# Optional: For advanced image processing
# numpy>=1.24.0,<2.0.0  # Uncomment if needed for advanced image processing

# Optional: For web scraping capabilities
# beautifulsoup4>=4.12.0,<5.0.0  # Uncomment if needed for content scraping

# Optional: For database storage
# sqlite3 (built-in)  # For local database storage if needed

# Optional: For email notifications
# smtplib (built-in)  # For email notifications if needed 